VCV-150: Refactor Review feature
