+++ src/features/review/site-detail/utils/normalize-form-data.ts
